WAKANDA GYM/requirements.txt
python-barcode==0.15.1
pillow>=9.0.0
requests>=2.25.0
mysql-connector-python>=8.0.0
python-dotenv>=0.19.0